backend/requirements.txt
fastapi
uvicorn[standard]
torch
torch-geometric
networkx
pandas
numpy
scikit-learn
openai
python-dotenv
scipy
